+++ porerefiner/jobs/__init__.py
@@ -11,23 +11,40 @@ from typing import Union, Tuple
 
 import logging
 import pkgutil
+import traceback
+
+JOBS = namedtuple('JOBS', ('FILES', 'RUNS'), defaults=([], []))() #configured (reified) job instances
+
+CLASS_REGISTRY = {} # available job classes
+
+CONFIGURED_JOB_REGISTRY = {}
 
 log = logging.getLogger('porerefiner.job')
 
 async def poll_active_job(job):
-    logg = log.getChild(f"{type(job.job_state.submitter).__name__}.{type(job.job_state).__name__}")
+    logg = log.getChild(f"poll")
+    configured_job = CONFIGURED_JOB_REGISTRY[job.job_class]
+    submitter = configured_job.submitter
     try:
-        await job.job_state.submitter._poll(job)
+        await submitter._poll(job)
     except Exception as e:
-        logg.error(e)
+        logg.error(f"error in {type(submitter).__name__} {type(configured_job).__name__} ")
+        # logg.error(e)
+        # traceback.print_exc()
+        raise
     return 1
 
 async def submit_job(job):
-    logg = log.getChild(f"{type(job.job_state.submitter).__name__}.{type(job.job_state).__name__}")
+    logg = log.getChild(f"submit")
+    configured_job = CONFIGURED_JOB_REGISTRY[job.job_class]
+    submitter = configured_job.submitter
     try:
-        await job.job_state.submitter._submit(job)
+        await submitter._submit(job)
     except Exception as e:
-        logg.error(e)
+        logg.error(f"error in {type(submitter).__name__} {type(configured_job).__name__} ")
+        # logg.error(e)
+        # traceback.print_exc()
+        raise
     return 1
 
 async def complete_job(job):
@@ -47,40 +64,57 @@ async def poll_jobs(ready_jobs, running_jobs):
     jobs_collected = sum(await gather(*[poll_active_job(job) for job in running_jobs]))
     return jobs_submitted + jobs_collected, jobs_submitted, jobs_collected
 
-JOBS = namedtuple('JOBS', ('FILES', 'RUNS'), defaults=([], []))() #configured (reified) job instances
 
-REGISTRY = {} # available job classes
+
 
 class _MetaRegistry(type):
 
     def __new__(meta, name, bases, class_dict):
+        "Register new CLASSES of job"
         cls = type.__new__(meta, name, bases, class_dict)
-        if cls not in REGISTRY:
-            REGISTRY[name] = cls
+        if cls not in CLASS_REGISTRY:
+            CLASS_REGISTRY[name] = cls
         return cls
 
+    # def __call__(cls, *args, **kwargs):
+    #     try:
+    #         the_instance = super().__call__(*args, **kwargs)
+    #     except TypeError as e:
+    #         raise TypeError(f"Bad config: {cls.__name__}: {e} ({args}, {kwargs})") from e
+    #     if isinstance(the_instance, FileJob):
+    #         JOBS.FILES.append(the_instance)
+    #         log.getChild('files').debug(cls.__name__)
+    #     if isinstance(the_instance, RunJob):
+    #         JOBS.RUNS.append(the_instance)
+    #         log.getChild('runs').debug(cls.__name__)
+    #     return the_instance
     def __call__(cls, *args, **kwargs):
+        "Register new INSTANCES of jobs"
         try:
             the_instance = super().__call__(*args, **kwargs)
         except TypeError as e:
             raise TypeError(f"Bad config: {cls.__name__}: {e} ({args}, {kwargs})") from e
+        except:
+            raise
         if isinstance(the_instance, FileJob):
             JOBS.FILES.append(the_instance)
             log.getChild('files').debug(cls.__name__)
         if isinstance(the_instance, RunJob):
             JOBS.RUNS.append(the_instance)
             log.getChild('runs').debug(cls.__name__)
+        CONFIGURED_JOB_REGISTRY[cls.__name__] = the_instance
+        assert the_instance
         return the_instance
 
 class RegisteringABCMeta(ABCMeta, _MetaRegistry):
+
     pass
+
 
 @dataclass
 class AbstractJob(metaclass=RegisteringABCMeta):
+
     submitter: Submitter
-    
-    def __post_init__(self):
-        self.attempts = 0
 
     @classmethod
     def get_configurable_options(cls):