+++ porerefiner/jobs/submitters/local.py
@@ -3,6 +3,8 @@ from dataclasses import dataclass
 from porerefiner.jobs.submitters import Submitter
 from asyncio import subprocess
 
+processes = {}
+
 
 @dataclass
 class LocalSubmitter(Submitter):
@@ -14,15 +16,15 @@ class LocalSubmitter(Submitter):
         return path #no op, no need to reroot paths
 
     async def begin_job(self, command, datadir, remotedir, environment_hints={}):
-        self.process = await subprocess.create_subprocess_shell(command)
+        processes[datadir] = await subprocess.create_subprocess_shell(command)
 
     async def poll_job(self, job):
-        await self.process.wait()
-        if self.process.returncode:
+        await processes[job.datadir].wait()
+        if processes[job.datadir].returncode:
             return 'FAILED'
         return 'DONE'
 
     def closeout_job(self, job, datadir, remotedir):
-        pass
+        del processes[datadir]